fix storage errors, sort of
Tauri seems to have an issue where permission to access the items in a folder doesn't persist after an app restart, even with the persistent storage plugin. I'll submit a bug for this but it's making it difficult to use this app, so I'm just going to grant full access to the Documents directory for now. This does mean that you can't store your markdown files outside of your system's Documents folder right now.